Cloud Functions/mongo-listener/requirements.txt
functions-framework==3.*
pymongo==4.6.1
google-cloud-storage==2.*